Leave components added
