+++ src/components/Leaves/Administrative.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Administrative() {
+  return <div></div>;
+}
+
+export default Administrative;
