+++ src/components/Leaves/AdministrativeFm.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function AdministrativeFm() {
+  return <div></div>;
+}
+
+export default AdministrativeFm;
