+++ src/components/Leaves/SickNoDoc.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function SickNoDoc() {
+  return <div></div>;
+}
+
+export default SickNoDoc;
