+++ src/components/Leaves/SickWithDoc.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function SickWithDoc() {
+  return <div></div>;
+}
+
+export default SickWithDoc;
