+++ src/components/Leaves/Social.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Social() {
+  return <div></div>;
+}
+
+export default Social;
