+++ src/components/Leaves/Study.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Study() {
+  return <div></div>;
+}
+
+export default Study;
