+++ src/components/Leaves/index.js
@@ -1,0 +1,7 @@
+export { default as Administrative } from './Administrative';
+export { default as AdministrativeFm } from './AdministrativeFm';
+export { default as Paid } from './Paid';
+export { default as SickNoDoc } from './SickNoDoc';
+export { default as SickWithDoc } from './SickWithDoc';
+export { default as Social } from './Social';
+export { default as Study } from './Study';
