+++ src/pages/NewRequest.js
@@ -20,21 +20,29 @@ import 'react-dates/lib/css/_datepicker.css';
 import { SingleDatePicker } from 'react-dates';
 import axios from 'axios';
 
+import {
+  Administrative,
+  AdministartiveFm,
+  Paid,
+  SickNoDoc,
+  SickWithDoc,
+  Social,
+  Study,
+} from '../components/Leaves/index';
+
 const leaves = [
-  'Paid leave',
-  'Administrative leave',
   'Administrative force majeure leave',
-  'Study leave',
+  'Administrative leave',
   'Social leave',
-  'Sick leave (with documents)',
   'Sick leave (no documents)',
+  'Sick leave (with documents)',
+  'Study leave',
+  'Paid leave',
 ];
 
-const roles = ['BA lead', 'QA', 'Junior'];
-const parts = ['Full', 'Fullest', 'Distance'];
-
-const am = ['Anna Marie', 'John Smith', 'Bill Anderson', 'Nill Armstrong', 'Susane Incredible'];
-const pm = [
+const prRoles = ['BA lead', 'QA', 'Junior'];
+const prParticipation = ['Full', 'Fullest', 'Distance'];
+const prManagers = [
   'Anney Kirillova',
   'Bill Ivanov',
   'Reygar Smith',
@@ -43,19 +51,8 @@ const pm = [
 ];
 
 function NewRequest({ isOpen, onClose }) {
-  const [leaveType, setLeaveType] = useState(0);
-  const [role, setRole] = useState(0);
-  const [part, setPart] = useState(0);
-  const [comment, setComment] = useState('');
-  const [fromDate, setFromDate] = useState(null);
-  const [focusedFrom, setFocusFrom] = useState(false);
-  const [focusedTo, setFocusTo] = useState(false);
-  const [toDate, setToDate] = useState(null);
-  const [amanager, setAManager] = useState([]);
-  const [pmanager, setPManager] = useState([]);
+  const [leaveType, setLeaveType] = useState(6);
   const [isSendingRequest, setRequestSending] = useState(false);
-
-  const getDateDifference = Math.round(toDate - fromDate) / (1000 * 60 * 60 * 24);
 
   const handleSendRequest = async () => {
     setRequestSending(true);
@@ -75,6 +72,29 @@ function NewRequest({ isOpen, onClose }) {
     }
     getAllData();
   }, []);
+
+  const renderLeaveBody = (type) => {
+    switch (type) {
+      case 0:
+        return (
+          <Administrative
+            prRoles={prRoles}
+            prParticipation={prParticipation}
+            prManagers={prManagers}
+            isSendingRequest={isSendingRequest}
+          />
+        );
+      case 6:
+        return (
+          <Paid
+            prRoles={prRoles}
+            prParticipation={prParticipation}
+            prManagers={prManagers}
+            isSendingRequest={isSendingRequest}
+          />
+        );
+    }
+  };
 
   return (
     <div>
@@ -102,168 +122,7 @@ function NewRequest({ isOpen, onClose }) {
             </Select>
           </FormControl>
 
-          <div
-            style={{
-              display: 'flex',
-              flexDirection: 'row',
-              marginBottom: 20,
-              alignItems: 'center',
-            }}>
-            <SingleDatePicker
-              disabled={isSendingRequest}
-              showClearDate
-              placeholder="From"
-              //transitionDuration={5000} WTF!
-              isDayBlocked={(day) => (toDate ? day > toDate : null)}
-              numberOfMonths={1}
-              date={fromDate}
-              onDateChange={(date) => setFromDate(date)}
-              focused={focusedFrom}
-              onFocusChange={({ focused }) => setFocusFrom(focused)}
-            />
-
-            <SingleDatePicker
-              disabled={isSendingRequest}
-              showClearDate
-              placeholder="To"
-              //transitionDuration={5} WTF!
-              isDayBlocked={(day) => (fromDate ? day < fromDate : null)}
-              numberOfMonths={1}
-              date={toDate}
-              onDateChange={(date) => setToDate(date)}
-              focused={focusedTo}
-              onFocusChange={({ focused }) => setFocusTo(focused)}
-            />
-
-            {fromDate && toDate && getDateDifference >= 0 ? (
-              <h4 style={{ paddingTop: 3 }}>
-                {getDateDifference + 1} {console.log(getDateDifference)} дней отпуска
-              </h4> //написать функцию склонения День
-            ) : null}
-          </div>
-
-          <div
-            style={{
-              display: 'flex',
-              flexDirection: 'column',
-              alignItems: 'center',
-            }}>
-            <TextField
-              disabled={isSendingRequest}
-              label="Comment"
-              variant="standard"
-              fullWidth
-              multiline
-              className="form-input"
-              value={comment}
-              onChange={(e) => {
-                setComment(e.target.value);
-              }}
-              style={{ marginBottom: 20, width: '100%' }}
-            />
-
-            <FormControl disabled={isSendingRequest} style={{ marginBottom: 20, width: '100%' }}>
-              <InputLabel>Project Role</InputLabel>
-              <Select
-                value={role}
-                onChange={(e) => {
-                  setRole(e.target.value);
-                }}>
-                {roles.map((item, idx) => (
-                  <MenuItem value={idx}>{item}</MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-
-            <FormControl disabled={isSendingRequest} style={{ marginBottom: 20, width: '100%' }}>
-              <InputLabel>Project participation</InputLabel>
-              <Select
-                value={part}
-                onChange={(e) => {
-                  setPart(e.target.value);
-                }}>
-                {parts.map((item, idx) => (
-                  <MenuItem value={idx}>{item}</MenuItem>
-                ))}
-              </Select>
-            </FormControl>
-          </div>
-
-          <div>
-            <h3>Approvers</h3>
-            <ol className="approvers__list">
-              <li>Accounting</li>
-              <li>
-                <FormControl disabled={isSendingRequest} className="approvers__item-form">
-                  <InputLabel id="am-label">AM</InputLabel>
-                  <Select
-                    labelId="am-label"
-                    multiple
-                    value={amanager}
-                    onChange={(e) => {
-                      setAManager(e.target.value);
-                    }}
-                    input={<Input />}
-                    //MenuProps={MenuProps}
-                  >
-                    {am.map((name) => (
-                      <MenuItem
-                        key={name}
-                        value={name}
-                        style={{ fontWeight: amanager.indexOf(name) === -1 ? '400' : '600' }}>
-                        {name}
-                      </MenuItem>
-                    ))}
-                  </Select>
-                </FormControl>
-                {amanager.length > 0 && (
-                  <Tooltip title="clear">
-                    <IconButton
-                      disabled={isSendingRequest}
-                      className="clear-icon"
-                      onClick={() => setAManager([])}>
-                      <CloseIcon />
-                    </IconButton>
-                  </Tooltip>
-                )}
-              </li>
-              <li>
-                <FormControl disabled={isSendingRequest} className="approvers__item-form">
-                  <InputLabel id="pm-label">PM</InputLabel>
-                  <Select
-                    labelId="pm-label"
-                    multiple
-                    value={pmanager}
-                    onChange={(e) => {
-                      setPManager(e.target.value);
-                    }}
-                    input={<Input />}
-                    //MenuProps={MenuProps}
-                  >
-                    {pm.map((name) => (
-                      <MenuItem
-                        key={name}
-                        value={name}
-                        style={{ fontWeight: pmanager.indexOf(name) === -1 ? '400' : '600' }}>
-                        {name}
-                      </MenuItem>
-                    ))}
-                  </Select>
-                </FormControl>
-                {pmanager.length > 0 && (
-                  <Tooltip title="clear">
-                    <IconButton
-                      disabled={isSendingRequest}
-                      className="clear-icon"
-                      onClick={() => setPManager([])}>
-                      <CloseIcon />
-                    </IconButton>
-                  </Tooltip>
-                )}
-              </li>
-              <li>CEO</li>
-            </ol>
-          </div>
+          {renderLeaveBody(leaveType)}
         </DialogContent>
         <DialogActions>
           <Button
